+++ .gitignore
@@ -88,3 +88,4 @@ __pycache__
 *.tmp
 test-results/
 yarn-error.log
+pkg/backend/display/wasm/mod.wasm
