+++ pkg/backend/display/wasm/main.go
@@ -1,31 +1,25 @@
-// Copyright 2016-2022, Pulumi Corporation.
+//go:build js && wasm
+
+// Copyright 2016-2023, Pulumi Corporation.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
 //
-//     http://www.apache.org/licenses/LICENSE-2.0
+//	http://www.apache.org/licenses/LICENSE-2.0
 //
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-
-//go:build !windows && !js
-// +build !windows,!js
-
-package rpcutil
+package main
 
 import (
-	"errors"
-	"os"
-
-	"github.com/pkg/term/termios"
+	_ "github.com/pulumi/pulumi/pkg/v3/backend/display"
 )
 
-var errUnsupported = errors.New("unsupported")
-
-func openPty() (*os.File, *os.File, error) {
-	return termios.Pty()
+func main() {
+	done := make(chan struct{}, 0)
+	<-done
 }
