+++ pkg/codegen/schema/loader_js.go
@@ -1,0 +1,12 @@
+//go:build js
+// +build js
+
+package schema
+
+import (
+	"time"
+)
+
+func (l *pluginLoader) loadCachedSchemaBytes(pkg string, path string, schemaTime time.Time) ([]byte, bool) {
+	return nil, false
+}
